Moved example recorded tests to /tests_not_to_run
