+++ src/main/java/tests_not_to_run/TestCase2.java
@@ -1,4 +1,4 @@
-package gov.state.history;
+package tests_not_to_run;
 
 import java.util.concurrent.TimeUnit;
 import org.junit.*;
@@ -6,7 +6,7 @@ import static org.junit.Assert.*;
 import org.openqa.selenium.*;
 import org.openqa.selenium.firefox.FirefoxDriver;
 
-public class TestCase1 {
+public class TestCase2 {
   private WebDriver driver;
   private String baseUrl;
   private boolean acceptNextAlert = true;
@@ -20,17 +20,14 @@ public class TestCase1 {
   }
 
   @Test
-  public void testCase1() throws Exception {
+  public void testCase2() throws Exception {
     driver.get(baseUrl + "/exist/apps/hsg-shell/");
     assertEquals("history.state.gov 3.0 shell", driver.getTitle());
-    driver.findElement(By.cssSelector("div.col-md-6 > ul > li > ul > li > a")).click();
-    assertEquals("Historical Documents", driver.findElement(By.cssSelector("h1")).getText());
-    assertEquals("Truman Administration 1945–1952", driver.findElement(By.cssSelector("td > a")).getText().replaceAll("\\r\\n|\\r|\\n", " "));
-    driver.findElement(By.cssSelector("td > a")).click();
-    assertEquals("TRUMAN ADMINISTRATION", driver.findElement(By.cssSelector("h1.titleStmt1")).getText());
-    assertEquals("General: The United Nations, Volume I", driver.findElement(By.cssSelector("ol > li > a > span")).getText());
-    driver.findElement(By.cssSelector("ol > li > a > span")).click();
-    assertEquals("Not found", driver.findElement(By.cssSelector("h1")).getText());
+    assertEquals("1750–1775", driver.findElement(By.xpath("(//a[contains(text(),'1750–1775')])[2]")).getText());
+    driver.findElement(By.xpath("(//a[contains(text(),'1750–1775')])[2]")).click();
+    assertEquals("Milestones: 1750–1775", driver.findElement(By.cssSelector("h1")).getText());
+    assertEquals("Rivalry in Europe, between the French and the British in particular, often influenced the course of events in their North American colonies.In an effort to increase their political and economic power, the British and the French competed to acquire the better share of the available land and control over the new trading opportunities the colonies presented.", driver.findElement(By.xpath("//div[@id='foreword']/p[2]")).getText());
+    assertEquals("Choose one 1750–1775 1776–1783 1784–1800 1801–1829 1830–1860 1861–1865 1866–1898 1899–1913 1914–1920 1921–1936 1937–1945 1945–1952 1953–1960 1961–1968 1969–1976 1977–1980 1981–1988 1989–1992 1993–2000 ... view all", driver.findElement(By.id("select-chapter")).getText());
   }
 
   @After
